easy_model_zoo/requirements.txt
opencv-python
torch==1.4.0
torchvision==0.5
Pillow
bounding-box==0.1.3
